% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description_v2.R
\name{build_OU_period_exp}
\alias{build_OU_period_exp}
\title{Generic function to build expactations for all datasets (number of DEs per dataset)}
\usage{
build_OU_period_exp(dataset_id, metadata_OrgUnit_datasetinfo, dataset_exp,
  period)
}
\arguments{
\item{dataset_id}{--> id of the dataset}

\item{metadata_OrgUnit_datasetinfo}{--> The dataframe containing orgunit_dataset metadata (should be the ouput of "extract_metadata_DS_OrgUnit" function)}

\item{dataset_exp}{--> The dataframe containing expactations for all datasets (should be the ouput of "build_DS_exp" function)}

\item{period}{--> should be a vector with a monthly range such as "c(201701, 201712)"}
}
\value{
Returns a dataframe with the number of expected values for each OrgUnit, dataset and period
}
\description{
Generic function to build expactations for all datasets (number of DEs per dataset)
}
