% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extracting_functions.R
\name{extract_categories}
\alias{extract_categories}
\title{Extract the categories for data elements}
\usage{
extract_categories(categories_url, userID, password)
}
\arguments{
\item{categories_url}{The url of the categories page in the DHIS api. The function is
made to parse xml pages, so input url should be an xml adress or a generic web
adress without extension.}

\item{userID}{your username in the given DHIS2 setting, as a character string}

\item{password}{your password for this DHIS2 setting, as a character string}
}
\value{
Returns a data frame with each category as a line and for each
category, its unique ID, its name and its url.
}
\description{
\code{extract_categories} extracts the list of categories that are used for different
data elements.
}

