% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract_wrapper.R
\name{extract_data}
\alias{extract_data}
\title{Extracting a data}
\usage{
extract_data(url_call, userID, password)
}
\arguments{
\item{url_call}{A data calling url as made by \link{make_extract_call}}

\item{userID}{your username in the given DHIS2 setting, as a character string}

\item{password}{your password for this DHIS2 setting, as a character string}
}
\value{
Returns a dataframe with one data value by line, and columns data_element_ID ,
period , org_unit_ID , value and category.
}
\description{
\code{extract_data} extracts data based on a url call
}

