% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract_content.R
\name{extract_dhis_content}
\alias{extract_dhis_content}
\title{Extract content information from DHIS}
\usage{
extract_dhis_content(base_url, userID, password)
}
\arguments{
\item{base_url}{The base url of the DHIS2 setting}

\item{userID}{your username in the given DHIS2 setting, as a character string}

\item{password}{your password for this DHIS2 setting, as a character string}
}
\value{
Returns a list of seven tables :

\strong{data_sets} The list of data sets as extracted by
\link{extract_dhis_datasets}.

\strong{data_elements} The list of data elements as extracted by
\link{extract_data_elements}.

\strong{data_elements_categories} The list of categories as extracted by
\link{extract_categories}.

\strong{org_units_list} The list of organization units as extracted by
\link{extract_orgunits_list}.

\strong{org_units_description} The description of organization units as extracted by
\link{extract_org_unit}.

\strong{org_units_group} The list of the groups of organization units as extracted by
\link{extract_org_unit}.

\strong{org_units_report} The list of reports for each organization unit as extracted by
\link{extract_org_unit}.
}
\description{
\code{extract_dhis_content} extracts content information from DHIS
}

