% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract_geolocalisation.R
\name{extract_geolocalisation}
\alias{extract_geolocalisation}
\title{Make shapefiles for organization units}
\usage{
extract_geolocalisation(org_units_description)
}
\arguments{
\item{org_units_description}{A data frame of organization descriptions as extracted with
\link{extract_org_unit}.}
}
\value{
Returns a list of two elements. The first element of the list is a shapefile of
point organization units. The second element is a shapefile of polygon organization
units.
}
\description{
\code{extract_geolocalisation} makes shapefiles from extracted organization units
}

