% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extracting_functions.R
\name{extract_info}
\alias{extract_info}
\title{Generic function to extract relevant nodes in DHIS element
\code{extract_info} goes to a specific DHIS2 element url, and extracts attributes
name , id and href. It can extract elements that span on multiple pages}
\usage{
extract_info(url_page, root, node_name, out, userID, password,
  monitor = FALSE)
}
\arguments{
\item{url_page}{The default url of the elemtns to parse in the DHIS api, as a
character string function is made to parse xml pages, so input url should be an xml
adress or a generic web adress without extension.}

\item{root}{root of this page, as extracted by \code{\link{parse_page}}}

\item{node_name}{the name of the name we wish to extract}

\item{out}{an empty dataframe in which to return the output (there are more elegant ways to do it for sure, see it later).}
}
\description{
Generic function to extract relevant nodes in DHIS element
\code{extract_info} goes to a specific DHIS2 element url, and extracts attributes
name , id and href. It can extract elements that span on multiple pages
}

