% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metadata.R
\name{extract_metadata_CC}
\alias{extract_metadata_CC}
\title{Generic function to extract Category combo meatadata from metadata list}
\usage{
extract_metadata_CC(list_metdata)
}
\arguments{
\item{list_metadata}{--> The list containing all metadata from a DHIS2.}
}
\value{
Returns a dataframe containing Category combo metadata
}
\description{
Generic function to extract Category combo meatadata from metadata list
}
