% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metadata.R
\name{extract_metadata_DS}
\alias{extract_metadata_DS}
\title{Generic function to extract DataSet meatadata from metadata list}
\usage{
extract_metadata_DS(list_metdata)
}
\arguments{
\item{list_metadata}{--> The list containing all metadata from a DHIS2.}
}
\value{
Returns a dataframe containing DataSet metadata
}
\description{
Generic function to extract DataSet meatadata from metadata list
}
