% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metadata.R
\name{extract_metadata_DS_OrgUnit}
\alias{extract_metadata_DS_OrgUnit}
\title{Generic function to show dataset available in Orgunits}
\usage{
extract_metadata_DS_OrgUnit(list_metdata, OrgUnit_pyr)
}
\arguments{
\item{OrgUnit_pyr}{--> The dataframe containing Orgunit pyramid (should be the ouput of "extract_metadata_OrgUnit" function)}

\item{list_metadata}{--> The list containing all metadata from a DHIS2}
}
\value{
Returns a dataframe containing Orgunit metadata related to datasets
}
\description{
Generic function to show dataset available in Orgunits
}
