% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metadata.R
\name{extract_metadata_OrgUnit}
\alias{extract_metadata_OrgUnit}
\title{Generic function to extract OrgUnit metadata from DHIS2 API}
\usage{
extract_metadata_OrgUnit(url, userID, password, list_metdata)
}
\arguments{
\item{url}{--> The url of the DHIS2 you want to extract data from, as a character string.}

\item{userID}{--> Your username in the given DHIS2 setting, as a character string}

\item{password}{--> Your password for this DHIS2 setting, as a character string}
}
\value{
Returns a dataframe containing Orgunit metadata
}
\description{
Generic function to extract OrgUnit metadata from DHIS2 API
}
