% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metadata.R
\name{extract_metdata}
\alias{extract_metdata}
\title{Generic function to extract json metadta file from DHIS2 API}
\usage{
extract_metdata(url, userID, password)
}
\arguments{
\item{url}{--> The url of the DHIS2 you want to extract data from, as a character string.}

\item{userID}{--> Your username in the given DHIS2 setting, as a character string}

\item{password}{--> Your password for this DHIS2 setting, as a character string}
}
\value{
Returns a list of metadata
}
\description{
Generic function to extract json metadta file from DHIS2 API
}
