% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extracting_functions.R
\name{extract_org_unit}
\alias{extract_org_unit}
\title{Extract information about an Orgunit}
\usage{
extract_org_unit(org_unit_url, userID, password)
}
\arguments{
\item{userID}{your username in the given DHIS2 setting, as a character string}

\item{password}{your password for this DHIS2 setting, as a character string}

\item{url}{The url of the organisation unit for which we want to extract
information. The function is made to parse xml pages, so input url should be an xml
adress or a generic web adress without extension.}
}
\value{
Returns a list with three elements :
* __Metadata__ For each organization unit, includes its geolocalization and
reference to parent unit

* __Group __ Groups in which the organization unit is included. This is where the
type of organization unit is stored

* __Datasets__ Datasets for which the organisation unit should communicate data
}
\description{
\code{extract_org_unit} extracts all the information about
}

