% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extracting_functions.R
\name{extract_orgunits_list}
\alias{extract_orgunits_list}
\title{Extract the list of Organisation Units in the DHIS setting}
\usage{
extract_orgunits_list(org_unit_page_url, userID, password)
}
\arguments{
\item{userID}{your username in the given DHIS2 setting, as a character string}

\item{password}{your password for this DHIS2 setting, as a character string}

\item{url}{The url of the organisation units page in the DHIS api. The function is
made to parse xml pages, so input url should be an xml adress or a generic web
adress without extension.}
}
\value{
Returns a data frame with each organisation unit as a line and for each
organisation unit, its unique ID, its name and its url.
}
\description{
\code{extract_orgunits_list} extracts the list of Organisation Units recorded in a
DHIS setting
}

