% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metadata.R
\name{flatten_hierarchy}
\alias{flatten_hierarchy}
\title{Generic function to show dataset available in Orgunits}
\usage{
flatten_hierarchy(metadata_OrgUnit, clean = TRUE)
}
\arguments{
\item{list_metadata}{--> The list containing all metadata from a DHIS2}

\item{OrgUnit_pyr}{--> The dataframe containing Orgunit pyramid (should be the ouput of "extract_metadata_OrgUnit" function)}
}
\value{
Returns a dataframe containing Orgunit metadata related to datasets
}
\description{
Generic function to show dataset available in Orgunits
}
