% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract_geolocalisation.R
\name{format_GPS}
\alias{format_GPS}
\title{Extracting GPS coordinates}
\usage{
format_GPS(org_units_description)
}
\arguments{
\item{org_units_description}{A data frame of organization descriptions as extracted with
\link{extract_org_unit}.}
}
\value{
Returns a list of data frames in which the columns are the organisation unit ID,
the longitude and the latitude of the organisation units. Each element of the list is an
organization unit.
}
\description{
\code{format_GPS} extracts GPS coordinates for extracted organization units
}

