% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_wrapper.R
\name{make_data_set_extract_call}
\alias{make_data_set_extract_call}
\title{Creating the data call adress}
\usage{
make_data_set_extract_call(base_url, data_sets, org_unit, period,
  update_date = "2009-01-01")
}
\arguments{
\item{base_url}{The base url of the DHIS2 setting}

\item{data_sets}{A table of data sets, as extracted}

\item{org_unit}{A table of organization units, as extracted}

\item{period_start}{Date of the beginning of the period from which to extract data}

\item{period_end}{Date of the end of the period from which to extract data}
}
\value{
Returns an url that calls on the data to be extracted based on inputted
parameters
}
\description{
\code{make__data_set_extract_call} creates a url used to call some data
}
