% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extracting_functions.R
\name{make_dhis_urls}
\alias{make_dhis_urls}
\title{Make relevant urls in DHIS web api}
\usage{
make_dhis_urls(dhis_url)
}
\arguments{
\item{url}{The url of the DHIS implementation}
}
\description{
\code{extract_org_unit} takes the main adress of a DHIS implementation and returns
the relevant adresses in the web api that will be used for extracting data.
}

