% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract_wrapper.R
\name{make_extract_call}
\alias{make_extract_call}
\title{Creating the data call adress}
\usage{
make_extract_call(base_url, data_sets, org_unit, period_start, period_end)
}
\arguments{
\item{base_url}{The base url of the DHIS2 setting}

\item{data_sets}{A table of data sets, as extracted by \link{extract_dhis_datasets}}

\item{org_unit}{A table of organization units, as extracted by \link{extract_org_unit}}

\item{period_start}{Date of the beginning of the period from which to extract data}

\item{period_end}{Date of the end of the period from which to extract data}
}
\value{
Returns an url that calls on the data to be extracted based on inputted
parameters
}
\description{
\code{make_extract_call} creates a url used to call some data
}

