% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract_geolocalisation.R
\name{make_shapefiles}
\alias{make_shapefiles}
\title{Make shapefiles for organization units}
\usage{
make_shapefiles(formatted_coordinates)
}
\arguments{
\item{formatted_coordinates}{A list of organization units and their GPS coordinates.
\link{format_GPS}.}
}
\value{
Returns a list of two elements. The first element of the list is a shapefile of
point organization units. The second element is a shapefile of polygon organization
units.
}
\description{
\code{make_shapefiles} makes shapefiles from a list of organization units coordinates.
}

