% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extracting_functions.R
\name{parse_page}
\alias{parse_page}
\title{Generic xml page parsing function}
\usage{
parse_page(url, userID, password)
}
\arguments{
\item{url}{The url of the page to parse in the DHIS api, as a character string. The
function is made to parse xml pages, so input url should be an xml adress or a
generic web adress without extension.}

\item{userID}{your username in the given DHIS2 setting, as a character string}

\item{password}{your password for this DHIS2 setting, as a character string}
}
\description{
Generic xml page parsing function
}

