% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_geolocalisation.R
\name{write_geolocalisation}
\alias{write_geolocalisation}
\title{Writing shapefiles for organization units}
\usage{
write_geolocalisation(shapefiles)
}
\arguments{
\item{shapefiles}{A list of organization units and their GPS coordinates extracted from \link{extract_geolocalisation}.}
}
\description{
\code{write_geolocalisation} makes and writes shapefiles from a list of organization units coordinates.
}
