% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extracting_functions.R
\name{extract_data_elements}
\alias{extract_data_elements}
\title{Extract the list of data elements in a DHIS data set}
\usage{
extract_data_elements(dataset_url, userID, password)
}
\arguments{
\item{userID}{your username in the given DHIS2 setting, as a character string}

\item{password}{your password for this DHIS2 setting, as a character string}

\item{url}{The url of the dataset page in the DHIS api, from which we want to
extract the data elements. The function is made to parse xml pages, so input url
should be an xml adress or a generic web adress without extension.}
}
\value{
Returns a data frame with each data element as a line and for each data
element, its unique ID, its name and its url.
}
\description{
\code{extract_data_elements} extracts the data elements recorded in a given dataset.
}

