% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extracting_functions.R
\name{extract_dhis_datasets}
\alias{extract_dhis_datasets}
\title{Extracting the list of datasets in DHIS}
\usage{
extract_dhis_datasets(url, userID, password)
}
\arguments{
\item{url}{The url of the datasets list in the DHIS web api, as a character string.
The function is made to parse xml pages, so input url should be an xml adress or a
generic web adress without extension.}

\item{userID}{your username in the given DHIS2 setting, as a character string}

\item{password}{your password for this DHIS2 setting, as a character string}
}
\value{
Returns a data frame with each dataset as a line and for each data set, its
unique ID, its name and its url.
}
\description{
\code{extract_dhis_datasets} goes to a specific DHIS2 implementation, and extracts
its full list of data sets
}

